backend/requirements.txt
fastapi 
uvicorn 
requests 
pdf2image 
pillow 
jwt
pytesseract 
python-Levenshtein
openai
python-multipart
python-docx
fuzzywuzzy
python-jose[cryptography] 
python-multipart
python-dotenv
openpyxl
fitz